+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-OAuth-authorization-server/src/main/java/com/graphql_maven_plugin/oauth/keycloak/RegularJsonConfigProviderFactory.java
@@ -1,6 +1,0 @@
-package com.graphql_maven_plugin.oauth.keycloak;
-
-import org.keycloak.services.util.JsonConfigProviderFactory;
-
-public class RegularJsonConfigProviderFactory extends JsonConfigProviderFactory {
-}