+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-OAuth-authorization-server/src/main/java/com/marcosbarbero/lab/sec/oauth/opaque/config/WebSecurityConfiguration.java
@@ -1,0 +1,57 @@
+package com.marcosbarbero.lab.sec.oauth.opaque.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.jdbc.JdbcDaoImpl;
+import org.springframework.security.crypto.factory.PasswordEncoderFactories;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+import javax.sql.DataSource;
+
+@EnableWebSecurity
+public class WebSecurityConfiguration extends WebSecurityConfigurerAdapter {
+
+    private final DataSource dataSource;
+
+    private PasswordEncoder passwordEncoder;
+    private UserDetailsService userDetailsService;
+
+    public WebSecurityConfiguration(final DataSource dataSource) {
+        this.dataSource = dataSource;
+    }
+
+    @Override
+    protected void configure(final AuthenticationManagerBuilder auth) throws Exception {
+        auth.userDetailsService(userDetailsService())
+                .passwordEncoder(passwordEncoder());
+    }
+
+    @Bean
+    @Override
+    public AuthenticationManager authenticationManagerBean() throws Exception {
+        return super.authenticationManagerBean();
+    }
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        if (passwordEncoder == null) {
+            passwordEncoder = PasswordEncoderFactories.createDelegatingPasswordEncoder();
+        }
+        return passwordEncoder;
+    }
+
+    @Bean
+    @Override
+    public UserDetailsService userDetailsService() {
+        if (userDetailsService == null) {
+            userDetailsService = new JdbcDaoImpl();
+            ((JdbcDaoImpl) userDetailsService).setDataSource(dataSource);
+        }
+        return userDetailsService;
+    }
+
+}