+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-OAuth-authorization-server/src/main/java/com/marcosbarbero/lab/sec/oauth/opaque/web/ProfileController.java
@@ -1,0 +1,19 @@
+package com.marcosbarbero.lab.sec.oauth.opaque.web;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.security.Principal;
+
+@RestController
+@RequestMapping("/profile")
+public class ProfileController {
+
+    @GetMapping("/me")
+    public ResponseEntity<Principal> get(final Principal principal) {
+        return ResponseEntity.ok(principal);
+    }
+
+}