+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-OAuth-authorization-server/src/main/java/org/allGraphQLCases/oauth2_authorization_server/package-info.java
@@ -1,1 +1,0 @@
-package org.allGraphQLCases.oauth2_authorization_server;